v1.1.2: overflow, parser and initialization fixes
Second correctness pass, aimed at the arithmetic and parsing defects this code
is prone to (see AGENTS.md).

- Integer overflow in size/offset arithmetic: TGA RLE write offsets, the HDR
  scanline size and the procedural mesh-generator vertex/index counts now
  compute in 64-bit / size_t and are range-checked before they truncate. The
  TGA BGR swap loop bound is 64-bit, so it can no longer wrap and skip the swap.
- Widened multiplication arithmetic: the flagged `count * stride` products (334
  sites across the decoders, mesh generators and math core) are computed in the
  wide type instead of being widened after a narrow 32-bit multiply, removing
  the associated signed-overflow undefined behaviour.
- Parser input: all numeric extraction in the Wavefront OBJ/MTL and Radiance HDR
  loaders moved from sscanf to strtof/strtol with endptr checking, so a
  malformed field is detectable rather than silently ignored.
- Uninitialized values: zero the glTF accessor unpack buffer before
  cgltf_accessor_unpack_floats (it may legitimately write fewer floats than
  requested) and the HDR RLE repeat state.
- Memory safety: explicit NULL guards around the Wavefront scene snapshot and
  the group-list append.
- Control flow: explicit default cases on switches over non-enum values.
- Portability: #pragma once replaced with include guards.
- Build: stb is pinned to a commit hash instead of the moving master ref, so a
  release is reproducible (AGENTS.md dependency licence gate).
- Added the cppcheck and clang-tidy analysis lanes and AGENTS.md.

Builds warning-free; clang-tidy and cppcheck report zero findings.

+++ src/glus_animation.c
@@ -75,14 +75,14 @@ GLUSvoid GLUSAPIENTRY glusAnimationSampleVec3f(GLUSfloat result[3], const GLUSfl
     /* Clamp after last keyframe. */
     if (t >= times[count - 1])
     {
-        glusVector3Copyf(result, values + (count - 1) * stride + vertexOffset);
+        glusVector3Copyf(result, values + (ptrdiff_t)(count - 1) * stride + vertexOffset);
         return;
     }
 
     i = animationFindSegment(times, count, t);
 
-    v0 = values + i * stride + vertexOffset;
-    v1 = values + (i + 1) * stride + vertexOffset;
+    v0 = values + (ptrdiff_t)i * stride + vertexOffset;
+    v1 = values + (ptrdiff_t)(i + 1) * stride + vertexOffset;
 
     if (interpolation == GLUS_ANIMATION_STEP)
     {
@@ -100,8 +100,8 @@ GLUSvoid GLUSAPIENTRY glusAnimationSampleVec3f(GLUSfloat result[3], const GLUSfl
 
     /* CUBICSPLINE: cubic Hermite spline per component. */
     delta = times[i + 1] - times[i];
-    m0    = values + i * stride + 6;   /* outTangent of keyframe i   */
-    m1    = values + (i + 1) * stride; /* inTangent of keyframe i+1  */
+    m0    = values + (ptrdiff_t)i * stride + 6;   /* outTangent of keyframe i   */
+    m1    = values + (ptrdiff_t)(i + 1) * stride; /* inTangent of keyframe i+1  */
 
     for (j = 0; j < 3; j++)
     {
@@ -141,14 +141,14 @@ GLUSvoid GLUSAPIENTRY glusAnimationSampleQuaternionf(GLUSfloat result[4], const 
     /* Clamp after last keyframe. */
     if (t >= times[count - 1])
     {
-        glusQuaternionCopyf(result, values + (count - 1) * stride + vertexOffset);
+        glusQuaternionCopyf(result, values + (ptrdiff_t)(count - 1) * stride + vertexOffset);
         return;
     }
 
     i = animationFindSegment(times, count, t);
 
-    v0 = values + i * stride + vertexOffset;
-    v1 = values + (i + 1) * stride + vertexOffset;
+    v0 = values + (ptrdiff_t)i * stride + vertexOffset;
+    v1 = values + (ptrdiff_t)(i + 1) * stride + vertexOffset;
 
     if (interpolation == GLUS_ANIMATION_STEP)
     {
@@ -180,8 +180,8 @@ GLUSvoid GLUSAPIENTRY glusAnimationSampleQuaternionf(GLUSfloat result[4], const 
 
     /* CUBICSPLINE: cubic Hermite spline per component, then normalize. */
     delta = times[i + 1] - times[i];
-    m0    = values + i * stride + 8;   /* outTangent of keyframe i   */
-    m1    = values + (i + 1) * stride; /* inTangent of keyframe i+1  */
+    m0    = values + (ptrdiff_t)i * stride + 8;   /* outTangent of keyframe i   */
+    m1    = values + (ptrdiff_t)(i + 1) * stride; /* inTangent of keyframe i+1  */
 
     for (j = 0; j < 4; j++)
     {
